outsource dailyTasks to db.js
