+++ src/db.js
@@ -1,0 +1,16 @@
+import {nanoid} from 'nanoid';
+const dailyTasks = [
+  {task: 'Räume lüften', isChecked: false, id: nanoid()},
+  {task: 'Betten machen', isChecked: false, id: nanoid()},
+  {task: 'Spülmaschine ausräumen', isChecked: false, id: nanoid()},
+  {task: 'Wäsche machen', isChecked: false, id: nanoid()},
+  {task: 'Müll rausbringen', isChecked: false, id: nanoid()},
+  {task: 'Böden saugen', isChecked: false, id: nanoid()},
+  {task: 'ungenutzte Dinge wegräumen', isChecked: false, id: nanoid()},
+  {task: 'Waschbecken im Bad reinigen', isChecked: false, id: nanoid()},
+  {task: 'Toilette im Bad reinigen', isChecked: false, id: nanoid()},
+  {task: 'Arbeitsflächen in der Küche abwischen', isChecked: false, id: nanoid()},
+  {task: 'volle Spülmaschine anschalten', isChecked: false, id: nanoid()},
+];
+
+export default dailyTasks;
